@@ -100,55 +100,56 @@
     "gulp-json-editor": "^2.5.5",
     "gulp-replace": "^1.1.0",
     "gulp-zip": "^5.1.0",
-    "html-loader": "^3.0.1",
+    "html-loader": "^4.1.0",
     "html-webpack-injector": "^1.1.4",
     "html-webpack-plugin": "^5.5.0",
-    "husky": "^7.0.4",
+    "husky": "^8.0.1",
     "jasmine-core": "^3.7.1",
     "jasmine-spec-reporter": "^7.0.0",
     "jest-mock-extended": "^2.0.6",
-    "jest-preset-angular": "^10.1.0",
-    "lint-staged": "^12.4.1",
+    "jest-preset-angular": "^12.1.0",
+    "lint-staged": "^13.0.3",
     "mini-css-extract-plugin": "^2.4.5",
     "node-ipc": "^9.2.1",
     "pkg": "5.7.0",
     "postcss": "^8.4.14",
-    "postcss-loader": "^6.2.1",
+    "postcss-loader": "^7.0.1",
     "prettier": "^2.6.2",
     "process": "^0.11.10",
     "regedit": "^3.0.3",
     "rimraf": "^3.0.2",
     "sass": "^1.34.1",
-    "sass-loader": "^12.4.0",
+    "sass-loader": "^13.0.2",
     "storybook-addon-designs": "^6.2.1",
     "style-loader": "^3.3.1",
     "tailwindcss": "^3.0.24",
     "tapable": "^1.1.3",
+    "ts-jest": "^28.0.6",
     "ts-loader": "^9.2.5",
     "tsconfig-paths-webpack-plugin": "^3.5.2",
-    "typescript": "4.3.5",
+    "typescript": "4.6.4",
     "url": "^0.11.0",
     "util": "^0.12.4",
     "webcrypto-shim": "^0.1.7",
     "webpack": "^5.64.4",
     "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "4.6",
+    "webpack-dev-server": "^4.9.3",
     "webpack-node-externals": "^3.0.0"
   },
   "dependencies": {
-    "@angular/animations": "^12.2.13",
-    "@angular/cdk": "^12.2.13",
-    "@angular/common": "^12.2.13",
-    "@angular/compiler": "^12.2.13",
-    "@angular/core": "^12.2.13",
-    "@angular/forms": "^12.2.13",
-    "@angular/platform-browser": "^12.2.13",
-    "@angular/platform-browser-dynamic": "^12.2.13",
-    "@angular/router": "^12.2.13",
+    "@angular/animations": "^14.0.6",
+    "@angular/cdk": "^14.0.4",
+    "@angular/common": "^14.0.6",
+    "@angular/compiler": "^14.0.6",
+    "@angular/core": "^14.0.6",
+    "@angular/forms": "^14.0.6",
+    "@angular/platform-browser": "^14.0.6",
+    "@angular/platform-browser-dynamic": "^14.0.6",
+    "@angular/router": "^14.0.6",
     "@koa/multer": "^3.0.0",
     "@koa/router": "^10.1.1",
-    "@microsoft/signalr": "^5.0.17",
-    "@microsoft/signalr-protocol-msgpack": "^5.0.17",
+    "@microsoft/signalr": "^6.0.7",
+    "@microsoft/signalr-protocol-msgpack": "^6.0.7",
     "big-integer": "^1.6.51",
     "bootstrap": "4.6.0",
     "braintree-web-drop-in": "^1.33.1",
