@@ -29,8 +29,8 @@ export class BrowserFileDownloadService implements FileDownloadService {
         true
       );
     } else {
-      if (navigator.msSaveOrOpenBlob) {
-        navigator.msSaveBlob(builder.blob, request.fileName);
+      if ((navigator as any).msSaveOrOpenBlob) {
+        (navigator as any).msSaveBlob(builder.blob, request.fileName);
       } else {
         const a = window.document.createElement("a");
         a.href = URL.createObjectURL(builder.blob);
